Updated arg parser to handle spaces
Parser now looks for quotations at beginning or end of "arguments"
to handle spaces in possible argument values. Beginning or end
quotations that are not preceded or followed (respectively) by a
space are ignored. if a quotation is missing, the parser returns an
empty tuple and prints an error message.